no need to start with very large test files
