+++ playwright/session-recording-network-recorder.spec.ts
@@ -1,0 +1,5 @@
+import { test } from './utils/posthog-playwright-test-base'
+
+test.describe('Session recording - array.js', () => {
+    test.fixme('network capture', () => {})
+})
